refactor(analysis-types): extract baseline file entry DTO
